+++ CHANGELOG.md
@@ -1,4 +1,9 @@
-## v0.0.1
+## v0.2
+ - Added tests on components and actions
+ - Added Sentiment Component
+ 
+ ## v0.1
  - Added tests on components and reducers
- - Added Cloud Component
+ - Added WordCloud omponent
+ - Added Cloud container omponent
  - Added app skeleton